+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Curse of Walking now increases max cost by 5 instead of 3
 * Heavily reduced swift sneak cost
+* Fixed enchant_with_treasure always giving all the enchantments instead of treasure only 
 
 ## 1.10.0
 * Added Gravity defying